requirements.txt
google-cloud-storage>=2.16.0
Pillow>=10.0.0
